MPSDataType TYPE_ID -> u32

+++ src/mps.rs
@@ -556,7 +556,7 @@ pub struct MPSIntersectionDistancePrimitiveIndexCoordinates {
 /// See <https://developer.apple.com/documentation/metalperformanceshaders/mpsdatatype?language=objc>.
 pub trait MPSDataType: Clone + Copy + PartialEq + Eq + Debug + Hash {
     type Type: Default + Clone + Copy + PartialEq + Debug + Sized;
-    const TYPE_ID: NSUInteger;
+    const TYPE_ID: u32;
 
     /// See <https://developer.apple.com/documentation/metalperformanceshaders/4092019-mpssizeofmpsdatatype?language=objc>.
     const SIZE: NSUInteger = ((Self::TYPE_ID & 0xFFFF) >> 3) as NSUInteger;
@@ -567,24 +567,24 @@ pub trait MPSDataType: Clone + Copy + PartialEq + Eq + Debug + Hash {
 }
 
 /// A common bit for all floating point data types. Zero for integer types
-const MPS_FLOATBIT_ENCODING: NSUInteger = 0x10000000;
+const MPS_FLOATBIT_ENCODING: u32 = 0x10000000;
 /// A common bit for all complex point data types. Zero for integer types
-const MPS_COMPLEXBIT_ENCODING: NSUInteger = MPS_FLOATBIT_ENCODING | 0x01000000;
+const MPS_COMPLEXBIT_ENCODING: u32 = MPS_FLOATBIT_ENCODING | 0x01000000;
 /// A common bit for all signed data types
-const MPS_SIGNEDBIT_ENCODING: NSUInteger = 0x20000000;
+const MPS_SIGNEDBIT_ENCODING: u32 = 0x20000000;
 /// A common bit for all alternate encoding data types
-const MPS_ALTERNATE_ENCODING: NSUInteger = 0x80000000;
+const MPS_ALTERNATE_ENCODING: u32 = 0x80000000;
 /// A common bit for all normalized data types.
 /// If set, the value of the shall be interpreted as value / UNORM_TYPE_MAX
 /// Normalized values have range [0, 1.0] if unsigned and [-1,1] if signed.
 /// SNORM_TYPE_MIN is interpreted as SNORM_TYPE_MIN+1 per standard Metal rules.
-const MPS_NORMALIZEDBIT_ENCODING: NSUInteger = 0x40000000;
+const MPS_NORMALIZEDBIT_ENCODING: u32 = 0x40000000;
 
 macro_rules! mps_datatype_impl {
     ($dt:ident, $dt_ty:ty, $type_id:expr, $from_f64:expr, $to_f64:expr) => {
         impl MPSDataType for $dt {
             type Type = $dt_ty;
-            const TYPE_ID: NSUInteger = $type_id;
+            const TYPE_ID: u32 = $type_id;
 
             fn from_f64(v: f64) -> Self::Type {
                 $from_f64(v)
@@ -844,7 +844,7 @@ impl MatrixDescriptor {
         rows: NSUInteger,
         columns: NSUInteger,
         row_bytes: NSUInteger,
-        data_type: NSUInteger,
+        data_type: u32,
     ) -> Self {
         unsafe {
             msg_send![
@@ -878,7 +878,7 @@ impl MatrixDescriptor {
         }
     }
 
-    pub fn row_bytes_for_columns(columns: NSUInteger, data_type: NSUInteger) -> NSUInteger {
+    pub fn row_bytes_for_columns(columns: NSUInteger, data_type: u32) -> NSUInteger {
         unsafe {
             msg_send![
                 class!(MPSMatrixDescriptor),
